src/requirements.txt
django
django-simple-history
simplejson
selenium
sphinx
sphinx_rtd_theme
